Switch to elf crate
Moved some things around. Think the idea now is to load the ELF into the page table and then fill in the rest of the table with unused memory. Then set a stack pointer and return everything. I think from there I should have a lot of the memory work for running an ELF done.