getters and setters for threads
Somewhat odd behavior without getters and setters. The values end up being 0 inside the switch to function. I suppose this could be because the thread struct is public but the fields aren't specified as public. Either way it's kind of odd that I can reference it after setting it and get the corresponding value back, but it's 0 when the switch_to function references that field itself. Getters and setters seem to resolve this issue.