Troubleshooting context switching back to a user thread
Kernel threads switch fine. A user thread will run and when it yields, it will get back to the scheduler fine, but when the scheduler tries to start it again it breaks. I have a suspicion that this might be a result of the launch_thread function, but need to do more troubleshooting to determine definitively